Allow for adding widgets to structure elements
