+++ phoebusgen/v4/screen.py
@@ -39,7 +39,7 @@ from xml.dom import minidom
 import os
 from phoebusgen.v4.properties import HasPosition, HasBackgroundColor, HasMacros, HasName, HasGrid, HasActionsRulesAndScripts, PropertyBase
 from phoebusgen.v4.properties.types import OpenDisplayAction
-from phoebusgen.v4.widgets import Widget
+from phoebusgen.v4.widgets import Widget, WidgetContainer
 from phoebusgen.v4.utils import prettify_xml
 from collections.abc import Sequence
 from pathlib import Path
@@ -48,10 +48,9 @@ from typing import TypeVar
 from phoebusgen.v4.widgets.structure import EmbeddedDisplay
 
 
-WidgetT = TypeVar('WidgetT', bound=Widget)
-PropertyT = TypeVar('PropertyT', bound=PropertyBase)
 
-class Screen(HasPosition, HasBackgroundColor, HasMacros, HasName, HasGrid, HasActionsRulesAndScripts):
+
+class Screen(WidgetContainer, HasPosition, HasBackgroundColor, HasMacros, HasName, HasGrid, HasActionsRulesAndScripts):
     """ Phoebus Screen object that holds widgets and can be written to .bob file """
     def __init__(self, name: str | None = None, f_name: str | None = None) -> None:
         """
@@ -96,48 +95,6 @@ class Screen(HasPosition, HasBackgroundColor, HasMacros, HasName, HasGrid, HasAc
             reparse_xml.writexml(f, indent='  ', addindent='  ', newl='\n', encoding='UTF-8')
         return True
 
-    def get_widgets(self) -> list[Widget]:
-        widget_elems = self.root.findall('widget')
-        widgets = []
-        for elem in widget_elems:
-            elem_type = elem.attrib.get('type', None)
-            if elem_type is None:
-                raise ValueError(f"Unknown widget type for element: {prettify_xml(elem)}")
-            widget_cls = None
-            for cls in Widget.__subclasses__():
-                if cls._widget_type is not None and cls._widget_type.value == elem_type:
-                    widget_cls = cls
-                    break
-            if widget_cls is None:
-                raise ValueError(f"Unsupported widget type '{elem_type}' for element: {prettify_xml(elem)}")
-            widgets.append(widget_cls.from_element(elem))
-        return widgets
-
-    def get_widgets_by_type(self, widget_type: type[WidgetT]) -> list[WidgetT]:
-        return [w for w in self.get_widgets() if isinstance(w, widget_type)]
-
-    def get_widgets_by_property_class(self, prop_type: type[PropertyT]) -> list[PropertyT]:
-        return [w for w in self.get_widgets() if isinstance(w, prop_type)]
-
-    def get_widgets_by_property(self, property_name: str) -> list[Widget]:
-        widgets_with_property = []
-        for widget in self.get_widgets():
-            if hasattr(widget, property_name):
-                widgets_with_property.append(widget)
-        return widgets_with_property
-
-    def add_widget(self, elem: Widget | Sequence[Widget]) -> None:
-        """
-        Add widget or list of widgets to screen
-
-        :param elem: <list/Phoebusgen.widget> List of Phoebusgen.widget's or a single widget to add
-        """
-        if isinstance(elem, Sequence):
-            for e in elem:
-                self.root.append(e.root)
-        else:
-            self.root.append(elem.root)
-
     def get_linked_screens(self) -> dict[Path, dict[str, str]]:
         linked_screens = {}
 
@@ -147,12 +104,16 @@ class Screen(HasPosition, HasBackgroundColor, HasMacros, HasName, HasGrid, HasAc
                 linked_screens[Path(action.file)] = copy.deepcopy(action.macros).update(self.macros)
 
         # Add any actions that open displays from the widgets
-        for widget in self.get_widgets():
-            for action in widget.actions:
-                if isinstance(action, OpenDisplayAction):
-                    linked_screens[Path(action.file)] = copy.deepcopy(action.macros).update(self.macros)
-                    if isinstance(widget, HasMacros):
-                        linked_screens[Path(action.file)].update(widget.macros)
+        def _get_actions_from_widgets(widget: Widget) -> Sequence[PropertyBase]:
+            for widget in self.get_widgets():
+                for action in widget.actions:
+                    if isinstance(action, OpenDisplayAction):
+                        linked_screens[Path(action.file)] = copy.deepcopy(action.macros).update(self.macros)
+                        if isinstance(widget, HasMacros):
+                            linked_screens[Path(action.file)].update(widget.macros)
+
+        for widget_container in [self, *self.get_widgets_by_type(WidgetContainer)]:
+            _get_actions_from_widgets(widget_container)
 
         # Add any embedded displays from the widgets
         for widget in self.get_widgets_by_type(EmbeddedDisplay):
